restructured and updatePriceCall
