+++ app/buy/interactions.tsx
@@ -1,8 +1,8 @@
 import { BrowserProvider, Contract, ethers, formatUnits } from 'ethers';
-import { optionFactoryABI } from '@/web3/OptionFactoryABI';
-import { callOptionABI } from '@/web3/CallOptionABI';
-import { putOptionABI } from '@/web3/PutOptionABI';
-import { erc20ABI } from '@/web3/ERC20ABI';
+import { optionFactoryABI } from '@/web3/abi/OptionFactoryABI';
+import { callOptionABI } from '@/web3/abi/CallOptionABI';
+import { putOptionABI } from '@/web3/abi/PutOptionABI';
+import { erc20ABI } from '@/web3/abi/ERC20ABI';
 import { optionFactory, noteAddress, tokenMapping } from '@/web3/constants';
 
 export interface OptionData {
