+++ app/positions/interactions.tsx
@@ -1,8 +1,8 @@
 import { BrowserProvider, Contract, ethers, formatUnits } from 'ethers';
-import { callOptionABI } from '@/web3/CallOptionABI';
-import { optionFactoryABI } from '@/web3/OptionFactoryABI';
-import { putOptionABI } from '@/web3/PutOptionABI';
-import { erc20ABI } from '@/web3/ERC20ABI';
+import { callOptionABI } from '@/web3/abi/CallOptionABI';
+import { optionFactoryABI } from '@/web3/abi/OptionFactoryABI';
+import { putOptionABI } from '@/web3/abi/PutOptionABI';
+import { erc20ABI } from '@/web3/abi/ERC20ABI';
 import { formatTimestamp } from '../buy/interactions';
 import { optionFactory, noteAddress, tokenMapping } from '@/web3/constants';
