+++ app/write/[token]/interactions.tsx
@@ -1,8 +1,8 @@
 import { BrowserProvider, Contract, ethers, parseEther } from 'ethers';
-import { optionFactoryABI } from '@/web3/OptionFactoryABI';
-import { erc20ABI } from '@/web3/ERC20ABI';
-import { putOptionABI } from '@/web3/PutOptionABI';
-import { callOptionABI } from '@/web3/CallOptionABI';
+import { optionFactoryABI } from '@/web3/abi/OptionFactoryABI';
+import { erc20ABI } from '@/web3/abi/ERC20ABI';
+import { putOptionABI } from '@/web3/abi/PutOptionABI';
+import { callOptionABI } from '@/web3/abi/CallOptionABI';
 import { optionFactory, noteAddress, reverseTokenMapping } from '@/web3/constants';
 
 export const createOptionCall = async (formData : any, walletProvider : any, chainId : any) => {
